Refactor proj structure
- split into useful modules, now trivial but will be populated later
- fix some typings for 3.10
