+++ apis/python/ocptv/dut.py
@@ -1,0 +1,3 @@
+class DutInfo:
+    def __str__(self):
+        return "dut_info_here"
